+++ cli/run.go
@@ -5,36 +5,32 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"os/exec"
-	"strings"
 
 	"github.com/johnewart/subcommands"
+	"github.com/yourbase/yb/internal/biome"
+	"github.com/yourbase/yb/internal/build"
 	"github.com/yourbase/yb/internal/ybdata"
 	"zombiezen.com/go/log"
 )
 
 type RunCmd struct {
-	environment string
+	target      string
+	noContainer bool
 }
 
 func (*RunCmd) Name() string     { return "run" }
 func (*RunCmd) Synopsis() string { return "Run an arbitrary command" }
 func (*RunCmd) Usage() string {
-	return `run [-e ENVIRONMENT] <COMMAND>
+	return `run [-t TARGET] <COMMAND>
 Run a command in the project container.
 `
 }
 
 func (p *RunCmd) SetFlags(f *flag.FlagSet) {
-	f.StringVar(&p.environment, "e", "default", "The environment to set")
+	f.StringVar(&p.target, "t", "default", "The target to run the command in")
+	f.BoolVar(&p.noContainer, "no-container", false, "Avoid using Docker if possible")
 }
 
-/*
-Executing the target involves:
-1. Map source into the target container
-2. Run any dependent components
-3. Start target
-*/
 func (b *RunCmd) Execute(ctx context.Context, f *flag.FlagSet, _ ...interface{}) subcommands.ExitStatus {
 	if len(f.Args()) == 0 {
 		fmt.Println(b.Usage())
@@ -45,53 +41,79 @@ func (b *RunCmd) Execute(ctx context.Context, f *flag.FlagSet, _ ...interface{})
 		log.Errorf(ctx, "%v", err)
 		return subcommands.ExitFailure
 	}
+	dockerClient, err := connectDockerClient(!b.noContainer)
+	if err != nil {
+		log.Errorf(ctx, "%v", err)
+		return subcommands.ExitFailure
+	}
+	pkg, err := GetTargetPackage()
+	if err != nil {
+		log.Errorf(ctx, "%v", err)
+		return subcommands.ExitFailure
+	}
+	targets, err := pkg.Manifest.BuildOrder(b.target)
+	if err != nil {
+		log.Errorf(ctx, "%v", err)
+		return subcommands.ExitFailure
+	}
+	dockerNetworkID, removeNetwork, err := newDockerNetwork(ctx, dockerClient)
+	if err != nil {
+		log.Errorf(ctx, "%v", err)
+		return subcommands.ExitFailure
+	}
+	defer removeNetwork()
 
-	targetPackage, err := GetTargetPackage()
+	// Build dependencies before running command.
+	err = doTargetList(ctx, pkg, targets[:len(targets)-1], &doOptions{
+		dockerClient:    dockerClient,
+		dockerNetworkID: dockerNetworkID,
+		dataDirs:        dataDirs,
+	})
 	if err != nil {
 		log.Errorf(ctx, "%v", err)
 		return subcommands.ExitFailure
 	}
 
-	instructions := targetPackage.Manifest
-
-	log.Infof(ctx, "Setting up dependencies...")
-	if err := targetPackage.SetupRuntimeDependencies(ctx, dataDirs); err != nil {
+	// Run command.
+	execTarget := targets[len(targets)-1]
+	bio, err := newBiome(ctx, dockerClient, pkg.Path)
+	if err != nil {
 		log.Errorf(ctx, "%v", err)
 		return subcommands.ExitFailure
 	}
-
-	log.Infof(ctx, "Setting environment variables...")
-	for _, property := range instructions.Exec.Environment["default"] {
-		s := strings.Split(property, "=")
-		if len(s) == 2 {
-			log.Infof(ctx, "  %s", s[0])
-			os.Setenv(s[0], s[1])
-		}
+	sys := build.Sys{
+		Biome:           bio,
+		DockerClient:    dockerClient,
+		DockerNetworkID: dockerNetworkID,
+		Stdout:          os.Stdout,
+		Stderr:          os.Stderr,
 	}
-
-	if b.environment != "default" {
-		for _, property := range instructions.Exec.Environment[b.environment] {
-			s := strings.Split(property, "=")
-			if len(s) == 2 {
-				log.Infof(ctx, "  %s", s[0])
-				os.Setenv(s[0], s[1])
-			}
-		}
+	phaseDeps, err := targetToPhaseDeps(execTarget)
+	if err != nil {
+		log.Errorf(ctx, "%v", err)
+		return subcommands.ExitFailure
 	}
-
-	execDir, _ := os.Getwd()
-	//execDir := filepath.Join(workspace.Path, targetPackage)
-
-	log.Infof(ctx, "Running %s from %s", strings.Join(f.Args(), " "), execDir)
-	cmdName := f.Args()[0]
-	cmdArgs := f.Args()[1:]
-	cmd := exec.Command(cmdName, cmdArgs...)
-	cmd.Dir = execDir
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-
-	cmd.Run()
-
+	execBiome, err := build.Setup(ctx, sys, phaseDeps)
+	if err != nil {
+		log.Errorf(ctx, "%v", err)
+		return subcommands.ExitFailure
+	}
+	// TODO(ch2744): Move this into build.Setup.
+	err = pkg.SetupBuildDependencies(ctx, dataDirs, execTarget)
+	if err != nil {
+		log.Errorf(ctx, "%v", err)
+		return subcommands.ExitFailure
+	}
+	// TODO(ch2725): Run the command from the subdirectory the process is in.
+	err = execBiome.Run(ctx, &biome.Invocation{
+		Argv:   f.Args(),
+		Stdin:  os.Stdin,
+		Stdout: os.Stdout,
+		Stderr: os.Stderr,
+	})
+	if err != nil {
+		log.Errorf(ctx, "%v", err)
+		return subcommands.ExitFailure
+	}
 	return subcommands.ExitSuccess
 }